Ajax

+++ ASP.NET.WebForm/ASP.NET.WebForm/CodeAdmin/Users.aspx.cs
@@ -9,133 +9,38 @@ using System.Data.SqlClient;
 using System.Configuration;
 using System.Data;
 using System.Windows.Forms;
-using BO = BussinesObject;
+using BussinesObject;
 using DAL = DataAccessLayer;
 using BAL = BusinessAccessLayer;
+
+
 namespace ASP.NET.WebForm.CodeAdmin
 {
     public partial class Users : System.Web.UI.Page
     {
         BAL.BAL_User uBAL = new BAL.BAL_User();
-        BO.BO_User User = new BO.BO_User();
-        SqlCommand cmd;
+        BO_User User = new BO_User();        
         SqlConnection conn = new SqlConnection(ConfigurationManager.ConnectionStrings["connString"].ConnectionString);        
-        SqlDataReader reader;
-        SqlDataAdapter adp;
         string query = string.Empty;
         codeSnippets ms = new codeSnippets();
         int search = 0;
         int create = 0;
         int update = 0;
         int delete = 0;
+        codeSnippets clear = new codeSnippets();
         protected void Page_Load(object sender, EventArgs e)
         {
             panelMsg.Visible = false;
-
-            #region Call whit Method & SP
-            //DataSet datos = new DataSet();
-            //try
-            //{
-            //    datos = TodoslsRegistros();
-            //    if (datos.Tables.Count > 0 && datos.Tables[0].Rows.Count > 0)
-            //    {
-            //        GridView4.DataSource = datos;
-            //        GridView4.DataBind();
-            //    }
-            //    else
-            //    {
-            //        GridView4.EmptyDataText = "No records Found";
-            //        GridView4.DataSource = null;
-            //        GridView4.DataBind();
-            //    }
-            //}
-            //catch (Exception ex)
-            //{
-            //    ms.MsgBox(ex.ToString(), Page, this);
-            //}
-            #endregion
-
-            #region Call whit SP
-            //DataSet ds = new DataSet();
-            //try
-            //{
-            //    ds = GetAllRecords();
-            //    if (ds.Tables.Count > 0 && ds.Tables[0].Rows.Count > 0)
-            //    {
-            //        GridView2.DataSource = ds;
-            //        GridView2.DataBind();
-            //    }
-            //    else
-            //    {
-            //        GridView2.EmptyDataText = "No records Found";
-            //        GridView2.DataSource = null;
-            //        GridView2.DataBind();
-            //    }
-            //}
-            //catch (Exception ex)
-            //{
-            //    ms.MsgBox(ex.ToString(), Page, this);
-            //}
-            #endregion
-
-            #region Call whit query
-            //query = "select * from dbo.SCRUD_RegisterUser";
-            //if (conn.State == ConnectionState.Closed)
-            //{
-            //    try
-            //    {
-            //        conn.Open();
-
-            //        cmd = new SqlCommand(query, conn);
-            //        reader = cmd.ExecuteReader();
-            //        while (reader.Read())
-            //        {
-            //            GridView3.DataSource = reader;
-            //            GridView3.DataBind();
-            //        }
-            //        reader.Close();
-            //        cmd.Dispose();
-            //        conn.Close();
-            //    }
-            //    catch (SqlException ex)
-            //    {
-            //        GridView3.EmptyDataText = "<b>No records Found</b>";
-            //        GridView3.DataSource = null;
-            //        GridView3.DataBind();
-            //        Response.Write("Failed: " + ex.Message);                    
-            //    }
-            //}
-            #endregion
+            //divUpdate.Visible = false;
 
             #region Four-Tier
             if (!IsPostBack)
                 RecordsGridView();
-            else
-                RecordsGridView();
+            //else
+            //    RecordsGridView();
             #endregion
         }
-        public DataSet GetAllRecords()
-        {
-            DataSet ds = new DataSet();
-            try
-            {
-                cmd = new SqlCommand("[SP_FetchRegisterUser]", conn);
-                cmd.CommandType = CommandType.StoredProcedure;
-                adp = new SqlDataAdapter(cmd);
-                adp.Fill(ds);
-                cmd.Dispose();
-            }
-            catch (Exception ex)
-            {
-                //throw new Exception("Upss! Error =( ", ex);
-                ms.MsgBox(ex.ToString(), Page, this);
-            }
-            finally
-            {
-                ds.Dispose();
-            }
-            return ds;
-        }
+
         public DataSet TodoslsRegistros()
         {            
             DataSet DS = new DataSet();
@@ -145,7 +50,7 @@ namespace ASP.NET.WebForm.CodeAdmin
             }
             catch (Exception ex)
             {
-                ms.MsgBox(ex.ToString(), Page, this);
+                throw ex;                
             }
             finally
             {
@@ -153,42 +58,44 @@ namespace ASP.NET.WebForm.CodeAdmin
             }
             return DS;
         }
-
-        //private void BindGrid()
-        //{
-        //    GridViewfourTier.DataSource = TodoslsRegistros();
-        //    GridViewfourTier.DataBind();
-        //}
-        
-        //GridViewSearch
         private void SearchRecords()
         {
-            User.Name = textSearch.Text.Trim();            
             DataTable dt = new DataTable();
-            try
-            {                
-                dt = uBAL.SearchBAL(User);                
-                if ( dt.Rows.Count > 0)
-                {
-                    GridViewSearch.DataSource = dt;
-                    GridViewSearch.DataBind();
-                }
-                else
-                {
-                    GridViewSearch.EmptyDataText = "No records Found";
-                    GridViewSearch.DataSource = null;
-                    GridViewSearch.DataBind();
-                }
-            }
-            catch (Exception ex)
+
+            if (codeSnippets.esLetra(textSearch.Text.Trim()) == true)
             {
-                Response.Write("Error: " + ex);
+                User.Name = textSearch.Text.Trim();
+                try
+                {
+                    dt = uBAL.SearchBAL(User);
+                    if (dt.Rows.Count > 0)
+                    {
+                        GridViewSearch.DataSource = dt;
+                        GridViewSearch.DataBind();
+                    }
+                    else
+                    {
+                        GridViewSearch.EmptyDataText = "<b>No existen registros!</b>";
+                        GridViewSearch.DataSource = null;
+                        GridViewSearch.DataBind();
+                    }
+                }
+                catch (Exception ex)
+                {
+                    Response.Write("Error: " + ex);
+                }
+                finally
+                {
+                    User = null;
+                    uBAL = null;
+                    codeSnippets.ClearTextBox(Page.Controls);
+                }        
             }
-            finally
+            else
             {
-                User = null;
-                uBAL = null;
-            }        
+                Response.Write(@"<script language='javascript'>alert('El campo solo acepta caracteres alfabeticos')</script>");
+                codeSnippets.ClearTextBox(Page.Controls);                
+            }                        
         }            
         protected void btnSearch_Click(object sender, EventArgs e)
         {
@@ -223,38 +130,80 @@ namespace ASP.NET.WebForm.CodeAdmin
         }
         protected void btnCreate_Click(object sender, EventArgs e)
         {
-            User.Name = textName.Text.Trim();
-            User.LastName = textApellidos.Text.Trim();
-            User.Email = textEmail.Text.Trim();
-            User.Company = textCompany.Text.Trim();
-            User.Date = textDate.Text.Trim();
-            User.Status = "1";                        
-
-            try
+            if (codeSnippets.esAlfanumerico(textName.Text.Trim()) == true)
             {
-                create = uBAL.CreateBAL(User);
-                if (create > 0)
+                User.Name = textName.Text.Trim();
+                if (codeSnippets.esAlfanumerico(textApellidos.Text.Trim()) == true)
                 {
-                    panelMsg.Visible = true;
-                    lblMsg.Text = "New record inserted successfully.";
-                    //ms.MsgBox("New record inserted successfully.", Page, this);                    
-                    //MessageBox.Show("New record inserted successfully.");
+                    User.LastName = textApellidos.Text.Trim();
+                    if (codeSnippets.esCorreo(textEmail.Text.Trim()) == true)
+                    {
+                        User.Email = textEmail.Text.Trim();
+                        if (codeSnippets.esAlfanumerico(textCompany.Text.Trim()) == true)
+                        {
+                            User.Company = textCompany.Text.Trim();
+                            User.Date = textDate.Text.Trim();
+                            User.Status = "1";
+                            
+                            try
+                            {
+                                create = uBAL.CreateBAL(User);
+                                if (create > 0)
+                                {
+                                    panelMsg.Visible = true;
+                                    lblMsg.Text = "New record inserted successfully.";
+                                    codeSnippets.ClearTextBox(Page.Controls);                
+                                }
+                                else
+                                    lblMsg.Text = "Cannot record inserted.";
+                            }
+                            catch (Exception ex)
+                            {
+                                Response.Write("Error: " + ex);
+                            }
+                            finally
+                            {
+                                User = null;
+                                uBAL = null;
+                            }                
+                        }
+                        else
+                        {
+                            Response.Write(@"<script language='javascript'>alert('Ingresa el nombre de la Compañia')</script>");                            
+                        }
+                    }
+                    else
+                    {
+                        Response.Write(@"<script language='javascript'>alert('Ingresa el Correo')</script>");
+                        //codeSnippets.ClearTextBox(Page.Controls);                
+                    }
                 }
-                else
-                    lblMsg.Text = "Cannot record inserted.";
+                else 
+                {
+                    Response.Write(@"<script language='javascript'>alert('Ingresa los Apellidos')</script>");
+                    //codeSnippets.ClearTextBox(Page.Controls);                
+                }
             }
-            catch (Exception ex)
+            else
             {
-                Response.Write("Error: "+ ex);                
-            }
-            finally
-            {
-                User = null;
-                uBAL = null;
-            }
+                Response.Write(@"<script language='javascript'>alert('Ingresa tu Nombre')</script>");
+                //codeSnippets.ClearTextBox(Page.Controls);                
+            }                     
         }
         protected void btnUpdate_Click(object sender, EventArgs e)
         {
+            DataSet ds = TodoslsRegistros();
+            //foreach (DataRow dr in ds.Tables[0].Rows)
+            //{
+
+
+            //    textUName.Text = dr["Name"].ToString();
+            //    textULastName.Text = dr["LastName"].ToString();
+            //    textUEmail.Text = dr["Email"].ToString();
+            //    textUCompany.Text = dr["Company"].ToString();
+            //    textUDate.Text = dr["Date"].ToString();
+            //}            
+            
             User.ID_User = Convert.ToInt16(testUID.Text.Trim());
             User.Name = textUName.Text.Trim();
             User.LastName = textULastName.Text.Trim();
@@ -262,27 +211,46 @@ namespace ASP.NET.WebForm.CodeAdmin
             User.Company = textUCompany.Text.Trim();
             User.Date = textUDate.Text.Trim();
             User.Status = "1";
-
-            try
+            
+            if (testUID.Text.Trim() != "")
             {
-                update = uBAL.UpdateBAL(User);
+                //divUpdate.Visible = true;
+                try
+                {
+                    update = uBAL.UpdateBAL(User);
+                }
+                catch (Exception ex)
+                {
+                    Response.Write("Error: " + ex);
+                }
+                finally
+                {
+                    User = null;
+                    uBAL = null;                    
+                }
             }
-            catch (Exception ex)
-            {
-                Response.Write("Error: " + ex);
-            }
-            finally
-            {
-                User = null;
-                uBAL = null;
-            }
+            //clear.ClearTextBox(Page.Controls);
         }  
         protected void btnDelete_Click(object sender, EventArgs e)
         {
+            //Validations from Code Behind 
+            if (codeSnippets.esNumero(textID.Text.Trim()) == true)
+            {                
             User.ID_User = Convert.ToInt16(textID.Text.Trim());
             try
-            {
-                delete = uBAL.DeleteBAL(User);
+            {                                               
+                DialogResult dr = MessageBox.Show("Are you sure you want to delete?", "Delete!", MessageBoxButtons.OKCancel, MessageBoxIcon.Information);
+                if (dr == DialogResult.OK)
+                {
+                    delete = uBAL.DeleteBAL(User);
+                    if (delete > 0)                    
+                        codeSnippets.ClearTextBox(Page.Controls);                    
+                    else                    
+                        //Response.Write(@"<script language='javascript'>alert('Ingresa el ID correcto')</script>");                        
+                        MessageBox.Show("Ese ID no existe, ingresa el ID correcto", "Advertencia", MessageBoxButtons.OK, MessageBoxIcon.Warning);                    
+                }
+                else                
+                    codeSnippets.ClearTextBox(Page.Controls);                
             }
             catch (Exception ex)
             {
@@ -293,6 +261,12 @@ namespace ASP.NET.WebForm.CodeAdmin
                 User = null;
                 uBAL = null;
             }
-        }      
+            }
+            else            
+                //Response.Write(@"<script language='javascript'>alert('El campo solo acepta Numeros')</script>");
+                MessageBox.Show("Ingresa solo datos Numericos", "Advertencia", MessageBoxButtons.OK, MessageBoxIcon.Warning);            
+            
+            //Ajax            
+        }                
     }
 }
